+++ examples/karma/src/index.js
@@ -1,0 +1,9 @@
+export const example = {
+  sum (a, b) {
+    return a + b;
+  },
+
+  subtract (a, b) {
+    return a - b;
+  }
+};